+++ airflow/contrib/operators/jenkins_job_trigger_operator.py
@@ -16,14 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.jenking.operators.jenkins_job_trigger`."""
+"""This module is deprecated. Please use `airflow.providers.jenkins.operators.jenkins_job_trigger`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.jenking.operators.jenkins_job_trigger import JenkinsJobTriggerOperator  # noqa
+from airflow.providers.jenkins.operators.jenkins_job_trigger import JenkinsJobTriggerOperator  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.jenking.operators.jenkins_job_trigger`.",
+    "This module is deprecated. Please use `airflow.providers.jenkins.operators.jenkins_job_trigger`.",
     DeprecationWarning, stacklevel=2
 )
