+++ docs/operators-and-hooks-ref.rst
@@ -929,8 +929,8 @@ These integrations allow you to perform various operations within various servic
 
    * - `Jenkins <https://jenkins.io/>`__
      -
-     - :mod:`airflow.providers.jenking.hooks.jenkins`
-     - :mod:`airflow.providers.jenking.operators.jenkins_job_trigger`
+     - :mod:`airflow.providers.jenkins.hooks.jenkins`
+     - :mod:`airflow.providers.jenkins.operators.jenkins_job_trigger`
      -
 
    * - `Opsgenie <https://www.opsgenie.com/>`__
