+++ tests/providers/jenkins/__init__.py
