+++ tests/providers/jenkins/operators/__init__.py
